Patch#11 Moving Type_handler_row related definitions into sql_type_row.cc
