+++ sql/sql_type.cc
@@ -214,50 +214,6 @@ bool Type_handler::is_traditional_scalar_type() const
 }
 
 
-class Type_collection_row: public Type_collection
-{
-public:
-  bool init(Type_handler_data *data) override
-  {
-    return false;
-  }
-  const Type_handler *aggregate_for_result(const Type_handler *a,
-                                           const Type_handler *b)
-                                           const override
-  {
-    return NULL;
-  }
-  const Type_handler *aggregate_for_comparison(const Type_handler *a,
-                                               const Type_handler *b)
-                                               const override
-  {
-    DBUG_ASSERT(a == &type_handler_row);
-    DBUG_ASSERT(b == &type_handler_row);
-    return &type_handler_row;
-  }
-  const Type_handler *aggregate_for_min_max(const Type_handler *a,
-                                            const Type_handler *b)
-                                            const override
-  {
-    return NULL;
-  }
-  const Type_handler *aggregate_for_num_op(const Type_handler *a,
-                                           const Type_handler *b)
-                                           const override
-  {
-    return NULL;
-  }
-};
-
-
-static Type_collection_row type_collection_row;
-
-const Type_collection *Type_handler_row::type_collection() const
-{
-  return &type_collection_row;
-}
-
-
 bool Type_handler_data::init()
 {
   return type_collection_geometry.init(this);
@@ -1762,11 +1718,6 @@ const Type_handler *Type_handler_timestamp_common::type_handler_for_comparison()
   return &type_handler_timestamp;
 }
 
-
-const Type_handler *Type_handler_row::type_handler_for_comparison() const
-{
-  return &type_handler_row;
-}
 
 /***************************************************************************/
 
@@ -3403,24 +3354,6 @@ bool Type_handler_bit::
 
 
 /*************************************************************************/
-bool Type_handler_row::Spvar_definition_with_complex_data_types(
-                                                 Spvar_definition *def) const
-{
-  if (def->row_field_definitions() && def->is_row())
-  {
-    List_iterator<Spvar_definition> it(*(def->row_field_definitions()));
-    Spvar_definition *member;
-    while ((member= it++))
-    {
-      if (member->type_handler()->is_complex())
-        return true;
-    }
-  }
-  return false;
-}
-
-
-/*************************************************************************/
 bool Type_handler::Key_part_spec_init_primary(Key_part_spec *part,
                                               const Column_definition &def,
                                               const handler *file) const
@@ -4434,12 +4367,6 @@ int Type_handler_bool::Item_save_in_field(Item *item, Field *field,
 
 /***********************************************************************/
 
-bool Type_handler_row::
-set_comparator_func(THD *thd, Arg_comparator *cmp) const
-{
-  return cmp->set_cmp_func_row(thd);
-}
-
 bool Type_handler_int_result::
 set_comparator_func(THD *thd, Arg_comparator *cmp) const
 {
@@ -4579,12 +4506,6 @@ bool Type_handler_numeric::
 
 
 /*************************************************************************/
-
-Item_cache *
-Type_handler_row::Item_get_cache(THD *thd, const Item *item) const
-{
-  return new (thd->mem_root) Item_cache_row(thd);
-}
 
 Item_cache *
 Type_handler_int_result::Item_get_cache(THD *thd, const Item *item) const
@@ -5915,12 +5836,6 @@ cmp_item *Type_handler_string_result::make_cmp_item(THD *thd,
   return new (thd->mem_root) cmp_item_sort_string(cs);
 }
 
-cmp_item *Type_handler_row::make_cmp_item(THD *thd,
-                                                    CHARSET_INFO *cs) const
-{
-  return new (thd->mem_root) cmp_item_row;
-}
-
 cmp_item *Type_handler_time_common::make_cmp_item(THD *thd,
                                                     CHARSET_INFO *cs) const
 {
@@ -6007,13 +5922,6 @@ Type_handler_timestamp_common::make_in_vector(THD *thd,
 }
 
 
-in_vector *Type_handler_row::make_in_vector(THD *thd,
-                                            const Item_func_in *func,
-                                            uint nargs) const
-{
-  return new (thd->mem_root) in_row(thd, nargs, 0);
-}
-
 /***************************************************************************/
 
 bool Type_handler_string_result::
@@ -6083,14 +5991,6 @@ bool Type_handler_temporal_result::
                                                     1U << (uint) TIME_RESULT);
 }
 
-
-bool Type_handler_row::Item_func_in_fix_comparator_compatible_types(THD *thd,
-                                              Item_func_in *func) const
-{
-  return func->compatible_types_row_bisection_possible() ?
-         func->fix_for_row_comparison_using_bisection(thd) :
-         func->fix_for_row_comparison_using_cmp_items(thd);
-}
 
 /***************************************************************************/
 
@@ -6359,32 +6259,6 @@ bool Type_handler_timestamp_common::
 }
 
 /***************************************************************************/
-
-/**
-  Get a string representation of the Item value.
-  See sql_type.h for details.
-*/
-String *Type_handler_row::
-          print_item_value(THD *thd, Item *item, String *str) const
-{
-  CHARSET_INFO *cs= thd->variables.character_set_client;
-  StringBuffer<STRING_BUFFER_USUAL_SIZE> val(cs);
-  str->append(STRING_WITH_LEN("ROW("));
-  for (uint i= 0 ; i < item->cols(); i++)
-  {
-    if (i > 0)
-      str->append(',');
-    Item *elem= item->element_index(i);
-    String *tmp= elem->type_handler()->print_item_value(thd, elem, &val);
-    if (tmp)
-      str->append(*tmp);
-    else
-      str->append(NULL_clex_str);
-  }
-  str->append(')');
-  return str;
-}
-
 
 /**
   Get a string representation of the Item value,
@@ -7767,38 +7641,6 @@ Item *Type_handler_temporal_with_date::
 }
 
 
-Item *Type_handler_row::
-  make_const_item_for_comparison(THD *thd, Item *item, const Item *cmp) const
-{
-  if (item->type() == Item::ROW_ITEM && cmp->type() == Item::ROW_ITEM)
-  {
-    /*
-      Substitute constants only in Item_row's. Don't affect other Items
-      with ROW_RESULT (eg Item_singlerow_subselect).
-
-      For such Items more optimal is to detect if it is constant and replace
-      it with Item_row. This would optimize queries like this:
-      SELECT * FROM t1 WHERE (a,b) = (SELECT a,b FROM t2 LIMIT 1);
-    */
-    Item_row *item_row= (Item_row*) item;
-    Item_row *comp_item_row= (Item_row*) cmp;
-    uint col;
-    /*
-      If item and comp_item are both Item_row's and have same number of cols
-      then process items in Item_row one by one.
-      We can't ignore NULL values here as this item may be used with <=>, in
-      which case NULL's are significant.
-    */
-    DBUG_ASSERT(item->result_type() == cmp->result_type());
-    DBUG_ASSERT(item_row->cols() == comp_item_row->cols());
-    col= item_row->cols();
-    while (col-- > 0)
-      resolve_const_item(thd, item_row->addr(col),
-                         comp_item_row->element_index(col));
-  }
-  return NULL;
-}
-
 /***************************************************************************/
 
 /*
@@ -8232,19 +8074,6 @@ void Type_handler_typelib::Item_param_set_param_func(Item_param *param,
 
 
 /***************************************************************************/
-
-Field *Type_handler_row::
-  make_table_field_from_def(TABLE_SHARE *share, MEM_ROOT *mem_root,
-                            const LEX_CSTRING *name,
-                            const Record_addr &rec, const Bit_addr &bit,
-                            const Column_definition_attributes *attr,
-                            uint32 flags) const
-{
-  DBUG_ASSERT(attr->length == 0);
-  DBUG_ASSERT(f_maybe_null(attr->pack_flag));
-  return new (mem_root) Field_row(rec.ptr(), name);
-}
-
 
 Field *Type_handler_olddecimal::
   make_table_field_from_def(TABLE_SHARE *share, MEM_ROOT *mem_root,
@@ -9736,18 +9565,4 @@ Item *Type_handler::make_typedef_constructor_item(THD *thd,
 {
   my_error(ER_WRONG_ARGUMENTS, MYF(0), def.get_name().str);
   return nullptr;
-}
-
-
-Item *Type_handler_row::make_typedef_constructor_item(THD *thd,
-                                                      const sp_type_def &def,
-                                                      List<Item> *args) const
-{
-  if (unlikely(args == nullptr))
-  {
-    my_error(ER_WRONG_ARGUMENTS, MYF(0), def.get_name().str);
-    return nullptr;
-  }
-
-  return new (thd->mem_root) Item_row(thd, *args);
 }
